fix: guard against nil in blake2b hash creation
This adds a guard that we will never hit, in practice. The blake2b.New
function can return a nil on error, so it triggers scanners like Uber's
nilaway. Adding a guard and panic here clears the alert, while
effectively changing nothing since we're always creating an empty hash
and that will always succeed.

Signed-off-by: Chris Gianelloni <wolf31o2@blinklabs.io>